# api/openapi.yaml
openapi: 3.1.0
info:
  title: FluxMQ Admin API
  description: |
    Admin-facing HTTP endpoints for FluxMQ broker monitoring and management.

    FluxMQ exposes two HTTP servers with distinct responsibilities:

    - **API Server** (`api_addr`): Admin endpoints for stats, cluster info, session
      management, and a combined overview. Also serves the Queue Service gRPC/Connect API.
    - **Health Server** (`health_addr`): Lightweight k8s probes for liveness, readiness,
      and basic cluster status.

    Both servers are internal/trusted — no authentication is required.
  version: 0.1.0
  license:
    name: Apache 2.0
    identifier: Apache-2.0
servers:
  - url: http://localhost:8080
    description: API Server (api_addr) — admin endpoints, sessions, queue service
  - url: http://localhost:8888
    description: Health Server (health_addr) — k8s liveness/readiness probes
...
paths:
...
  /api/v1/stats:
    get:
      operationId: getStats
      summary: Broker statistics
      description: |
        Returns all broker counters in one call. Every field is an atomic load —
        zero allocation, zero contention. Raw cumulative counters are exposed;
        clients compute rates by polling.
      tags: [Admin]
      servers:
        - url: http://localhost:8080
          description: API Server
      responses:
        "200":
          description: Current broker statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResponse"
        "405":
          description: Method not allowed (non-GET request)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Broker not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: broker not available

    # ── Stats (aggregated + per-protocol) ──
    ConnectionStats:
      type: object
      properties:
        current:
          type: integer
          format: uint64
          description: Currently connected clients (sum across protocols)
          example: 342
        total:
          type: integer
          format: uint64
          description: Total connections since startup
          example: 15000
        disconnections:
          type: integer
          format: uint64
          description: Total disconnections since startup
          example: 14658
    MessageStats:
      type: object
      description: Aggregated message counters across all protocols
      properties:
        received:
          type: integer
          format: uint64
          example: 1250000
        sent:
          type: integer
          format: uint64
          example: 980000
    ByteStats:
      type: object
      properties:
        received:
          type: integer
          format: uint64
          example: 524288000
        sent:
          type: integer
          format: uint64
          example: 412000000
    ErrorStats:
      type: object
      description: Aggregated protocol error count
      properties:
        protocol:
          type: integer
          format: uint64
          example: 12
    MQTTMessageStats:
      type: object
      properties:
        received:
          type: integer
          format: uint64
          example: 1100000
        sent:
          type: integer
          format: uint64
          example: 850000
        publish_received:
          type: integer
          format: uint64
          description: PUBLISH packets received
          example: 1000000
        publish_sent:
          type: integer
          format: uint64
          description: PUBLISH packets sent
          example: 800000
    MQTTSubscriptionStats:
      type: object
      properties:
        active:
          type: integer
          format: uint64
          description: Currently active MQTT subscriptions
          example: 1200
        retained_messages:
          type: integer
          format: uint64
          description: Number of retained messages stored
          example: 45
    MQTTErrorStats:
      type: object
      properties:
        protocol:
          type: integer
          format: uint64
          example: 12
        auth:
          type: integer
          format: uint64
          description: Authentication failures
          example: 5
        authz:
          type: integer
          format: uint64
          description: Authorization failures
          example: 3
        packet:
          type: integer
          format: uint64
          description: Malformed packet errors
          example: 8
    MQTTStats:
      type: object
      description: MQTT-specific counters
      properties:
        connections:
          $ref: "#/components/schemas/ConnectionStats"
        messages:
          $ref: "#/components/schemas/MQTTMessageStats"
        bytes:
          $ref: "#/components/schemas/ByteStats"
        subscriptions:
          $ref: "#/components/schemas/MQTTSubscriptionStats"
        errors:
          $ref: "#/components/schemas/MQTTErrorStats"
    AMQPStats:
      type: object
      description: AMQP 0.9.1-specific counters
      properties:
        connections:
          $ref: "#/components/schemas/ConnectionStats"
        messages:
          $ref: "#/components/schemas/MessageStats"
        bytes:
          $ref: "#/components/schemas/ByteStats"
        channels:
          type: integer
          format: uint64
          description: Currently open AMQP channels
          example: 42
        consumers:
          type: integer
          format: uint64
          description: Currently active AMQP consumers
          example: 15
        errors:
          $ref: "#/components/schemas/ErrorStats"
    ByProtocolStats:
      type: object
      properties:
        mqtt:
          $ref: "#/components/schemas/MQTTStats"
        amqp:
          $ref: "#/components/schemas/AMQPStats"
    StatsResponse:
      type: object
      description: |
        Top-level fields are aggregated totals across all protocols.
        The `by_protocol` section contains per-protocol breakdowns with
        protocol-specific fields (e.g. MQTT subscriptions, AMQP channels).
      properties:
        uptime_seconds:
          type: number
          format: double
          description: Broker uptime in seconds
          example: 86400.5
        connections:
          $ref: "#/components/schemas/ConnectionStats"
        messages:
          $ref: "#/components/schemas/MessageStats"
        bytes:
          $ref: "#/components/schemas/ByteStats"
        errors:
          $ref: "#/components/schemas/ErrorStats"
        by_protocol:
          $ref: "#/components/schemas/ByProtocolStats"

    # ── Cluster ──
...